add CompareSubSamples
