+++ src/Components/FilterBox.js
@@ -15,6 +15,8 @@ import CalendarParent from "../dashboard/CalendarParent";
 import RenderCheckContainer from "./FilterBoxComponents/RenderCheckContainer";
 import Grid from "@material-ui/core/Grid";
 
+import CompareSubSamples from "./CompareSubsamples";
+
 export default function FilterBox(props) {
   const History = useHistory();
   const {
@@ -240,57 +242,74 @@ export default function FilterBox(props) {
       //render Compare SubSamples if on BarChart
     } else if (filterSelectorName === "Compare SubSamples" && open === "bar") {
       return (
-        <div>
-          <form>
-            {/* labels filter */}
-            <p>{filterSelectorName}</p>
-            <Select
-              defaultValue={{ label: filters[index].selectedCategory }}
-              // isClearable
-              //seems not in use
-              formatGroupLabel={formatGroupLabel}
-              components={{ Control: ControlComponent }}
-              // isSearchable
-              onChange={e => {
-                setUpdateUrlFlag(!updateUrlFlag);
-                let optionFlags = {};
-                graphLabels[
-                  `${FilterBoxOptions.default[e.label].value.type}`
-                ].labels.forEach(option => {
-                  optionFlags = {
-                    ...optionFlags,
-                    [option]: false
-                  };
-                });
-                setFilters({
-                  ...filters,
-                  [index]: {
-                    ...filters[index],
-                    selectedCategory: e.label, //option
-                    selectedTableColumnName:
-                      FilterBoxOptions.default[e.label].value.type,
-
-                    selectedTable:
-                      FilterBoxOptions.default[e.label].value.query,
-                    selectedOption: undefined,
-                    selectableOptions: { ...optionFlags }
-                  }
-                });
-              }}
-              name="color"
-              styles={colourStyles}
-              options={xVar(
-                FilterBoxOptions.superCategories,
-                Object.keys(filters)
-                  .map(filterId => {
-                    return filters[filterId].selectedCategory;
-                  })
-                  .filter(selectedCategory => selectedCategory.length > 0)
-              )}
-            />
-          </form>
-        </div>
+        <CompareSubSamples
+          filterSelectorName={filterSelectorName}
+          filters={filters}
+          setFilters={setFilters}
+          formatGroupLabel={formatGroupLabel}
+          ControlComponent={ControlComponent}
+          index={index}
+          formatGroupLabel={formatGroupLabel}
+          setUpdateUrlFlag={setUpdateUrlFlag}
+          FilterBoxOptions={FilterBoxOptions}
+          updateUrlFlag={updateUrlFlag}
+          xVar={xVar}
+          colourStyles={colourOptions}
+        />
       );
+
+      // return (
+      //   <div>
+      //     <form>
+      //       {/* labels filter */}
+      //       <p>{filterSelectorName}</p>
+      //       <Select
+      //         defaultValue={{ label: filters[index].selectedCategory }}
+      //         // isClearable
+      //         //seems not in use
+      //         formatGroupLabel={formatGroupLabel}
+      //         components={{ Control: ControlComponent }}
+      //         // isSearchable
+      //         onChange={e => {
+      //           setUpdateUrlFlag(!updateUrlFlag);
+      //           let optionFlags = {};
+      //           graphLabels[
+      //             `${FilterBoxOptions.default[e.label].value.type}`
+      //           ].labels.forEach(option => {
+      //             optionFlags = {
+      //               ...optionFlags,
+      //               [option]: false
+      //             };
+      //           });
+      //           setFilters({
+      //             ...filters,
+      //             [index]: {
+      //               ...filters[index],
+      //               selectedCategory: e.label, //option
+      //               selectedTableColumnName:
+      //                 FilterBoxOptions.default[e.label].value.type,
+
+      //               selectedTable:
+      //                 FilterBoxOptions.default[e.label].value.query,
+      //               selectedOption: undefined,
+      //               selectableOptions: { ...optionFlags }
+      //             }
+      //           });
+      //         }}
+      //         name="color"
+      //         styles={colourStyles}
+      //         options={xVar(
+      //           FilterBoxOptions.superCategories,
+      //           Object.keys(filters)
+      //             .map(filterId => {
+      //               return filters[filterId].selectedCategory;
+      //             })
+      //             .filter(selectedCategory => selectedCategory.length > 0)
+      //         )}
+      //       />
+      //     </form>
+      //   </div>
+      // );
       //compare subsamples not to render on map or lineGraph
     } else if (filterSelectorName === "Compare SubSamples") {
       return <></>;
